requirements.txt
asgiref==3.4.1
Django==3.2.7
django-crispy-forms==1.13.0
djangorestframework==3.12.4
Pillow==8.3.2
pytz==2020.1
sqlparse==0.3.1
